pyproject.toml
[tool.poetry]
name = "django-oscar-poetry-fly"
version = "0.1.0"
description = ""
authors = ["Chef <cheftech@patraldo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-oscar = {extras = ["sorl-thumbnail"], version = "^3.2.5"}
whitenoise = "^6.7.0"
pillow = "^11.0.0"
django-widget-tweaks = "^1.5.0"
pycountry = "^24.6.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
